Revert "fix compilation error"
This reverts commit 88307a81921f78aa878a0b891ed0acff38b4a9e5.

+++ iter/entry_iterator_test.go
@@ -194,7 +194,7 @@ type generator func(i int64) logproto.Entry
 
 func mkStreamIterator(f generator, labels string) EntryIterator {
 	entries := []logproto.Entry{}
-	for i := range int64(testSize) {
+	for i := range testSize {
 		entries = append(entries, f(i))
 	}
 	return NewStreamIterator(logproto.Stream{
